Fix Truncation in Infolog
I have long been irritated by filenames and other things being truncated in infolog. I'm assuming I'm not the only one. Could this be increased to something like 200 or so to make truncation exceedingly rare?

Not a hill to die on or anything, just a minor (very occasionally major) annoyance.